point the generator to production and regenerated with the latest release - 18eb8aa
