+++ nile_api/api/workspaces/delete_access_token.py
@@ -1,0 +1,101 @@
+from typing import Any, Dict
+
+import httpx
+
+from ...client import Client
+from ...types import Response
+
+
+def _get_kwargs(
+    workspace: str,
+    id: str,
+    *,
+    client: Client,
+) -> Dict[str, Any]:
+    url = "{}/workspaces/{workspace}/access_tokens/{id}".format(
+        client.base_url, workspace=workspace, id=id
+    )
+
+    headers: Dict[str, str] = client.get_headers()
+    cookies: Dict[str, Any] = client.get_cookies()
+
+    return {
+        "method": "delete",
+        "url": url,
+        "headers": headers,
+        "cookies": cookies,
+        "timeout": client.get_timeout(),
+    }
+
+
+def _build_response(*, response: httpx.Response) -> Response[Any]:
+    return Response(
+        status_code=response.status_code,
+        content=response.content,
+        headers=response.headers,
+        parsed=None,
+    )
+
+
+def sync_detailed(
+    workspace: str,
+    id: str,
+    *,
+    client: Client,
+) -> Response[Any]:
+    """Delete an access token
+
+     Delete a workspace access token. Any users or services using this token will no longer be able to
+    access the workspace.
+
+    Args:
+        workspace (str):
+        id (str):
+
+    Returns:
+        Response[Any]
+    """
+
+    kwargs = _get_kwargs(
+        workspace=workspace,
+        id=id,
+        client=client,
+    )
+
+    response = httpx.request(
+        verify=client.verify_ssl,
+        **kwargs,
+    )
+
+    return _build_response(response=response)
+
+
+async def asyncio_detailed(
+    workspace: str,
+    id: str,
+    *,
+    client: Client,
+) -> Response[Any]:
+    """Delete an access token
+
+     Delete a workspace access token. Any users or services using this token will no longer be able to
+    access the workspace.
+
+    Args:
+        workspace (str):
+        id (str):
+
+    Returns:
+        Response[Any]
+    """
+
+    kwargs = _get_kwargs(
+        workspace=workspace,
+        id=id,
+        client=client,
+    )
+
+    async with httpx.AsyncClient(verify=client.verify_ssl) as _client:
+        response = await _client.request(**kwargs)
+
+    return _build_response(response=response)
